+++ R/shiny--csvInput.R
@@ -26,27 +26,8 @@ csvInputServer <- function(id, initial_df, sig_digits = 4) {
       shiny::req(input$file)
       file <- input$file$datapath
 
-      tryCatch(new_data <- readr::read_csv(file),
-               error = \(cnd) shiny::showModal(shiny::modalDialog(
-                 title = shiny::p(
-                   shiny::icon("ban"), "Error reading the uploaded file"
-                   ),
-                 csv_reading_error(),
-                 easyClose = TRUE
-               ))
-               )
-
-      tryCatch(assert_data.frame(new_data, template = initial_df),
-               error = \(cnd) shiny::showModal(shiny::modalDialog(
-                 title = shiny::p(
-                   shiny::icon("ban"), "Invalid data format"
-                   ),
-                 csv_format_error(),
-                 easyClose = TRUE
-               ))
-               )
-
-      shiny::req(FALSE)
+      new_data <- csv_try_read(file)
+      csv_try_assert_df(new_data, template = initial_df)
 
       df(new_data)
     })
@@ -67,8 +48,9 @@ csvInputServer <- function(id, initial_df, sig_digits = 4) {
   })
 }
 
-csv_reading_error <- function()
-  shiny::markdown(
+csv_try_read <- function(file) {
+  err_title <- shiny::p(shiny::icon("ban"), "Error reading the uploaded file")
+  err_msg <-   shiny::markdown(
     "The file could not be read as a valid CSV. Please ensure that the file:
 
     * Is in CSV format (e.g., .csv extension).
@@ -77,14 +59,39 @@ csv_reading_error <- function()
 
     Try re-uploading the file or using a different one."
   )
+  err_modal <- shiny::modalDialog(title = err_title, err_msg, easyClose = TRUE)
+  err_fn <- function(cnd) {
+    shiny::showModal(err_modal)
+    shiny::req(FALSE)
+  }
 
-csv_format_error <- function()
-  shiny::markdown(
-    "The uploaded file does not match the expected structure. Please ensure that:
+  tryCatch(readr::read_csv(file), error = err_fn)
+}
 
-    * The file contains the required columns: <list of expected column names> (case-sensitive).
-    * Each column has the correct data type: <column name> should be <expected type>.
-    * There are no missing or unexpected columns.
+csv_try_assert_df <- function(df, template) {
+  err_title <- shiny::p(shiny::icon("ban"), "Invalid data format")
 
-    Check the file and upload it again."
-  )
+  cols <- colnames(template)
+  types <- sapply(template, typeof)
+  cols_list <- paste(cols, collapse = ", ")
+  types_msg <- paste0("\n   * '", cols, "' should be of type '", types, "'") |>
+    paste0(collapse = "")
+  err_msg <- shiny::markdown(paste(
+    "The uploaded file does not match the expected structure.",
+    "Please ensure that:\n\n",
+    "* The file contains the required columns:", cols_list, "\n",
+    "* Each column has the correct data type:", types_msg, "\n\n",
+    "Check the file and upload it again."
+    ))
+
+  err_modal <- shiny::modalDialog(title = err_title, err_msg, easyClose = TRUE)
+  err_fn <- function(cnd) {
+    shiny::showModal(err_modal)
+    shiny::req(FALSE)
+  }
+
+  tryCatch(assert_data.frame(df, template = template), error = err_fn)
+}
+
+
+